services/instapy/requirements.txt
https://github.com/timgrossmann/InstaPy/zipball/gui-features
websocket-client
requests
python-dotenv
psutil